+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 4.17.0 (2026-07-06)
+
+Full Changelog: [v4.16.2...v4.17.0](https://github.com/trycourier/courier-go/compare/v4.16.2...v4.17.0)
+
+### Features
+
+* **openapi:** Journeys cancel-by-token endpoint + send-node experiments (C-18986) ([ae6187e](https://github.com/trycourier/courier-go/commit/ae6187e92fbd2548f2d1a52849cd693483a28c12))
+
+
+### Documentation
+
+* **openapi:** reword Cancel journey runs description (re-trigger deploy) ([98ef901](https://github.com/trycourier/courier-go/commit/98ef901c6b4b4b14796de2c136841344a721f897))
+* **openapi:** tighten Journeys cancel/experiment copy (C-19177) ([3fb48d3](https://github.com/trycourier/courier-go/commit/3fb48d3ffc536dd3c14ec4535acdc951390aad7f))
+
 ## 4.16.2 (2026-06-25)
 
 Full Changelog: [v4.16.1...v4.16.2](https://github.com/trycourier/courier-go/compare/v4.16.1...v4.16.2)